Uses Free Pascal from Linux distribution instead the upstream
Currently the project is downloading and building an upstream
version of Free Pascal. The script can use the Pascal compiler
provide from the Linux distribution instead.

Using the package manage tool from distribution, it is not
necessary to verify the architecture of the OS to use the
correct Free Pascal version. Package manage can resolve it
for the script.